fix: wrap search page useSearchParams in Suspense boundary
Next.js requires useSearchParams() to be wrapped in a Suspense boundary
at the page level for static rendering. Split search page into server
component (page.tsx) wrapping a client component (SearchRedirect.tsx).
